Moved list into new TodoList component
