+++ src/App.js
@@ -1,29 +1,11 @@
 import React from 'react';
-
-const todoList = [
-  {
-    id: 1,
-    title: 'Washing dishes'
-  },
-  {
-    id: 2,
-    title: 'Feeding cats'
-  },
-  {
-    id: 3,
-    title: 'Morning walk'
-  }
-];
+import TodoList from './TodoList';
 
 function App() {
   return (
     <div>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map(function (item) {
-          return <li key={item.id}>{item.title}</li>;
-        })}
-      </ul>
+      <TodoList />
     </div>
   );
 }
